feat: authorization is added

+++ src/controllers/tasks.ts
@@ -3,7 +3,6 @@ import Task from "../models/Task";
 
 // @desc    Get all tasks
 // @route   GET /api/v1/tasks
-// @access  Private/Admin
 export const getTasks = async (
   req: Request,
   res: Response,
@@ -23,7 +22,6 @@ export const getTasks = async (
 
 // @desc    Get single task
 // @route   GET /api/v1/tasks/:id
-// @access  Private/Admin
 export const getTask = async (
   req: Request,
   res: Response,
@@ -42,7 +40,6 @@ export const getTask = async (
 
 // @desc    Create task
 // @route   POST /api/v1/tasks
-// @access  Private/Admin
 export const createTask = async (
   req: Request,
   res: Response,
@@ -61,21 +58,44 @@ export const createTask = async (
 
 // @desc    Update task
 // @route   PUT /api/v1/tasks/:id
-// @access  Private/Admin
 export const updateTask = async (
   req: Request,
   res: Response,
   next: NextFunction
 ) => {
-  const task = await Task.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  try {
+    const taskId = req.params.id;
+    const updatedData = req.body;
 
-  res.status(200).json({
-    success: true,
-    data: task,
-  });
+    // Find the task by ID
+    const task = await Task.findById(taskId);
+
+    if (!task) {
+      return res.status(404).json({
+        success: false,
+        error: "Task not found",
+      });
+    }
+
+    // Check if the logged-in user is the task owner
+    if (task.user.toString() !== req.user?._id.toString()) {
+      return res.status(403).json({
+        success: false,
+        error: "You are not authorized to update this task",
+      });
+    }
+
+    // Update the task data
+    Object.assign(task, updatedData);
+    await task.save();
+
+    res.status(200).json({
+      success: true,
+      data: task,
+    });
+  } catch (error) {
+    return next(error);
+  }
 };
 
 // @desc    Delete task
@@ -86,17 +106,71 @@ export const deleteTask = async (
   next: NextFunction
 ) => {
   try {
-    const task = await Task.findById(req.params.id);
+    const taskId = req.params.id;
+
+    // Find the task by ID
+    const task = await Task.findById(taskId);
+
     if (!task) {
       return res.status(404).json({
         success: false,
-        error: "No task found",
+        error: "Task not found",
       });
     }
+
+    // Check if the logged-in user is the task owner
+    if (task.user.toString() !== req.user?._id.toString()) {
+      return res.status(403).json({
+        success: false,
+        error: "You are not authorized to delete this task",
+      });
+    }
+
     await task.deleteOne();
+
     return res.status(200).json({
       success: true,
       data: {},
+    });
+  } catch (error) {
+    return next(error);
+  }
+};
+
+// @desc    Mark task as complete
+// @route   PUT /api/v1/tasks/:id/complete
+export const markTaskAsComplete = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  try {
+    const taskId = req.params.id;
+
+    // Find the task by ID
+    const task = await Task.findById(taskId);
+
+    if (!task) {
+      return res.status(404).json({
+        success: false,
+        error: "Task not found",
+      });
+    }
+
+    // Check if the logged-in user is the task owner
+    if (task.user.toString() !== req.user?._id.toString()) {
+      return res.status(403).json({
+        success: false,
+        error: "You are not authorized to mark this task as complete",
+      });
+    }
+
+    task.status = "Done"; // Set task status as "Done"
+    await task.save();
+
+    return res.status(200).json({
+      success: true,
+      data: task,
     });
   } catch (error) {
     return next(error);
